refactored frontend
