+++ components/common/CardStat.tsx
@@ -1,9 +1,18 @@
 import ResponsiveTooltip from "@/components/common/Tooltip";
 
-export default function CardStat({ id, label, value, secondaryValue, children, tooltip }: { id: string, label: string, value?: string, secondaryValue?: string, children?: JSX.Element, tooltip: string }): JSX.Element {
+export interface CardStatProps {
+  id: string;
+  label: string;
+  value?: string;
+  secondaryValue?: string;
+  children?: JSX.Element;
+  tooltip: string;
+}
+
+export default function CardStat({ id, label, value, secondaryValue, children, tooltip }: CardStatProps): JSX.Element {
   return <div className="w-full md:w-1/4 flex flex-row md:block justify-between md:justify-normal mt-6 md:mt-0">
     <p
-      className="text-white font-normal md:text-sm w-1/2 md:w-full"
+      className="text-white font-normal w-1/2 md:w-full"
       id={id}
     >
       {label}
