+++ components/landing/Hero.tsx
@@ -1,7 +1,7 @@
-import StatusWithLabel from "@/components/common/StatusWithLabel";
 import { networthAtom, tvlAtom, vaultronAtom } from "@/lib/atoms";
 import { NumberFormatter } from "@/lib/utils/formatBigNumber";
 import { useAtom } from "jotai";
+import LargeCardStat from "@/components/common/LargeCardStat";
 
 export default function Hero(): JSX.Element {
   const [networth] = useAtom(networthAtom)
@@ -10,128 +10,60 @@ export default function Hero(): JSX.Element {
 
   return (
     <section className="pb-8 pt-8 sm:pb-6 border-b border-customGray300">
-      <div className="w-full flex flex-col smmd:flex-row smmd:items-center justify-between px-4 md:px-8">
-        <div className="w-full xl:w-8/12 md:flex md:flex-row space-y-4 md:space-y-0 mt-4 md:mt-0">
+      <div className="w-full px-4 md:px-8 md:flex flex-row items-center">
+        <div className="w-full xl:w-2/3 md:flex md:flex-row space-y-4 md:space-y-0 mt-4 md:mt-0">
           <div className="flex flex-row items-center md:pr-10 gap-10 md:w-fit">
-            <StatusWithLabel
-              label={"Deposits"}
-              content={
-                <p className="text-3xl font-bold text-white leading-[120%]">
-                  ${" "}
-                  {
-                    networth.vault > 0.01
-                      ? NumberFormatter.format(networth.vault)
-                      : "0"
-                  }
-                </p>
-              }
-              className="md:min-w-[160px] lg:min-w-0"
+            <LargeCardStat
+              id="networth-deposits"
+              label="Deposits"
+              value={`$ ${networth.vault > 0.01 ? NumberFormatter.format(networth.vault) : "0"}`}
+              tooltip=""
             />
-            <StatusWithLabel
-              label={"Staked"}
-              content={
-                <p className="text-3xl font-bold text-white leading-[120%]">
-                  ${" "}
-                  {
-                    networth.stake > 0.01
-                      ? NumberFormatter.format(networth.stake)
-                      : "0"
-                  }
-                </p>
-              }
-              className="md:min-w-[160px] lg:min-w-0"
+            <LargeCardStat
+              id="networth-staked"
+              label="Staked"
+              value={`$ ${networth.stake > 0.01 ? NumberFormatter.format(networth.stake) : "0"}`}
+              tooltip=""
             />
-
           </div>
 
           <div className="flex flex-row items-center md:pr-10 gap-10 md:w-fit">
-            <StatusWithLabel
-              label={"VCX in Wallet"}
-              content={
-                <p className="text-3xl font-bold text-white leading-[120%]">
-                  ${" "}
-                  {
-                    networth.wallet > 0.01
-                      ? NumberFormatter.format(networth.wallet)
-                      : "0"
-                  }
-                </p>
-              }
-              className="md:min-w-[160px] lg:min-w-0"
+            <LargeCardStat
+              id="networth-wallet"
+              label="Wallet"
+              value={`$ ${networth.wallet > 0.01 ? NumberFormatter.format(networth.wallet) : "0"}`}
+              tooltip=""
             />
-            <StatusWithLabel
-              label={"My Net Worth"}
-              content={
-                <p className="text-3xl font-bold text-white leading-[120%]">
-                  ${" "}
-                  {
-                    networth.total > 0.01
-                      ? NumberFormatter.format(networth.total)
-                      : "0"
-                  }
-                </p>
-              }
-              infoIconProps={{
-                id: "networth",
-                title: "My Networth",
-                content: (
-                  <p>
-                    This value aggregates your VaultCraft-related <br />{" "}
-                    holdings across all blockchain networks.
-                  </p>
-                ),
-              }}
-              className="md:min-w-[160px] lg:min-w-0"
+            <LargeCardStat
+              id="networth-total"
+              label="Net Worth"
+              value={`$ ${networth.stake > 0.01 ? NumberFormatter.format(networth.stake) : "0"}`}
+              tooltip=""
             />
           </div>
-          <StatusWithLabel
-            label={"My XP"}
-            content={
-              <p className="text-3xl font-bold text-white leading-[120%]">
-                {vaultronStats.xp}
-              </p>
-            }
-            infoIconProps={{
-              id: "vaultronXp",
-              title: "My XP",
-              content: (
-                <p>
-                  This value shows your Vaultron XP
-                </p>
-              ),
-            }}
-            className="md:min-w-[160px] lg:min-w-0"
-          />
+          <div className="flex flex-row items-center md:w-fit">
+            <LargeCardStat
+              id="networth-xp"
+              label="XP"
+              value={`${vaultronStats.xp} XP`}
+              tooltip=""
+            />
+          </div>
         </div>
 
-        <div className="w-full md:w-1/3 mt-8 md:mt-0">
+        <div className="w-full xl:w-1/3 mt-8 md:mt-0">
           <p className="uppercase md:hidden text-customGray200 text-sm mb-2">
             Platform
           </p>
           <div className="w-full flex flex-row items-center md:justify-end">
-            <StatusWithLabel
-              label={"Total Value Locked"}
-              content={
-                <p className="text-3xl font-bold text-white leading-[120%]">
-                  {
-                    tvl.total > 0.01
-                      ? NumberFormatter.format(tvl.total)
-                      : "0"
-                  }
-                </p>
-              }
-              infoIconProps={{
-                id: "tvl",
-                title: "Total Value Locked",
-                content: (
-                  <p>
-                    Total value locked (TVL) is the amount <br /> of user funds
-                    deposited in VaultCraft products.
-                  </p>
-                ),
-              }}
-              className="md:min-w-[160px]"
-            />
+            <div className="w-fit">
+              <LargeCardStat
+                id="tvl"
+                label="Total Value Locked"
+                value={`$ ${tvl.total > 0.01 ? NumberFormatter.format(tvl.total) : "0"}`}
+                tooltip=""
+              />
+            </div>
           </div>
         </div>
       </div>
