+++ components/landing/Products.tsx
@@ -134,18 +134,10 @@ export default function Products(): JSX.Element {
             description="Deposit and optimize your yield with automated, non-custodial DeFi strategies"
             stats={[
               {
+                id: "smart-vault-tvl",
                 label: "TVL",
-                content: `$${NumberFormatter.format(tvl.vault)}`,
-                infoIconProps: {
-                  title: "Total Value Locked",
-                  content: (
-                    <p>
-                      The total value of assets held <br /> by the underlying
-                      smart contracts.
-                    </p>
-                  ),
-                  id: "smart-vault-tvl",
-                },
+                value: `$${NumberFormatter.format(tvl.vault)}`,
+                tooltip: "The total value of assets held <br /> by the underlying smart contracts."
               },
             ]}
             route="vaults"
